% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_amazon.R
\name{amazon_bearer_tokens}
\alias{amazon_bearer_tokens}
\title{Bearer tokens to access Amazon API}
\usage{
amazon_bearer_tokens()
}
\value{
A string with the token
}
\description{
This function randomly samples a valid token to use the Fake Amazon API.
}
\examples{
amazon_bearer_tokens()
}
\author{
Jorge Cimentada
}
