% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_amazon.R
\name{api_amazon}
\alias{api_amazon}
\title{REST API for the Books/Products of Amazon Website}
\usage{
api_amazon()
}
\value{
callr result of the individual R process
}
\description{
Using the Faker API (https://fakerapi.it/en), this API generates fake but reasonably structured data representing books and products from an internal 'Amazon API'. This API serves as educational purposes only.
}
\details{
This function launches a plumber API in a new R process using `callr` and return the `callr` process.
}
\examples{

\dontrun{
  live_api <- api_amazon()
  live_api$kill()
}

}
\author{
Jorge Cimentada
}
