% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_bicing.R
\name{api_bicing}
\alias{api_bicing}
\title{REST API for Bicycle usage in Barcelona}
\usage{
api_bicing(port = NULL)
}
\arguments{
\item{port}{a numeric value used as a port}
}
\value{
callr result of the individual R process
}
\description{
Bicing is a Bicycle-sharing system in the city of Barcelona, Spain, which was implemented in March 2007, promoted by the City Council.
}
\details{
This API wraps a manually downloaded set of bicycle stations from the city. Every time the API is invoked, a random number of bicycles are assigned as 'being used' and the timestamp in the returning dataset is updated with the current time.

The API has no parameters

This function launches a plumber API in a new R process using `callr` and return the `callr` process.
}
\examples{

\dontrun{
  live_api <- api_bicing()
  live_api$kill()
}

}
\author{
Jorge Cimentada
}
