% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanish_schools_ex.R
\name{prep_browser}
\alias{prep_browser}
\title{Prepare local HTML links to open with your browser}
\usage{
prep_browser(x)
}
\arguments{
\item{x}{a character vector of HTML files}
}
\value{
the same character vector but with "file://" appended to it
}
\description{
Prepare local HTML links to open with your browser
}
\examples{

library(xml2)
library(scrapex)

one_link <- spanish_schools_ex()[5]
prep_browser(one_link)

\dontrun{
# You can open it with your browser with browseURL
browseURL(prep_browser(one_link))
}

}
\author{
Jorge Cimentada
}
