% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_coveragedb.R
\name{serve_api_coveragedb}
\alias{serve_api_coveragedb}
\title{Custom rest API for examples}
\usage{
serve_api_coveragedb(region, sex)
}
\value{
Whatever
}
\description{
Custom rest API for examples
}
\author{
Jorge Cimentada
}
