% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spanish_schools_ex.R
\name{spanish_schools_ex}
\alias{spanish_schools_ex}
\title{Return the paths to the complete HTML website of a sample of Spanish Schools}
\usage{
spanish_schools_ex()
}
\value{
A character vector with 27 local HTML links
}
\description{
This example hosts data from the website https://www.buscocolegio.com/
which tries to aggregate data on all schools in Spain. To build reliable
and self-standing scraping examples, this package downloaded data on 27
randomly sampled schools from this website. This function returns the full
HTML links to the local HTML files that can be used as exercises for
scraping some data.
}
\details{
All images should work well out of the box except the Google maps on the
right. Please note that these websites are not meant
to be used to reliably explore their hyperlink or other features, as
these will change inevitably when the website changes some of it's
structures. This is because some of the links still point to the
real website.
}
\examples{

library(xml2)
library(scrapex)

one_link <- spanish_schools_ex()[1]

read_html(one_link)

\dontrun{
# If you want to look at the website in your browser
# just wrap it in prep_browser and pass it to browseURL
browseURL(prep_browser(one_link))
}

}
\author{
Jorge Cimentada
}
