% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_coveragedb.R
\name{api_coveragedb}
\alias{api_coveragedb}
\title{Custom rest API for examples}
\usage{
api_coveragedb()
}
\value{
Whatever
}
\description{
Custom rest API for examples
}
\examples{

\dontrun{
  ho()
}

}
\author{
Jorge Cimentada
}
