% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{launch_api}
\alias{launch_api}
\title{Launches an API in a new R process}
\usage{
launch_api(function_name, random_port = NULL)
}
\arguments{
\item{function_name}{the folder name where the plumber API is located. This should be in inst/plumber/}

\item{random_port}{a numeric value used as a port}
}
\value{
callr result of the individual R process
}
\description{
Launches an API in a new R process
}
\examples{

\dontrun{
  live_api <- launch_api("api_coveragedb")
  live_api$kill()
}

}
\author{
Jorge Cimentada
}
